Fix test - refactor folder
